fix: [DHIS2-9119] [DHIS2-9120] Integer out of ranges for long ids (2.33)

+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/system/objectmapper/OrganisationUnitRelationshipRowMapper.java
@@ -44,7 +44,7 @@ public class OrganisationUnitRelationshipRowMapper
     public OrganisationUnitRelationship mapRow( ResultSet resultSet )
         throws SQLException
     {
-        return new OrganisationUnitRelationship( resultSet.getInt( "parentid" ), resultSet.getInt( "organisationunitid" ) );
+        return new OrganisationUnitRelationship( resultSet.getLong( "parentid" ), resultSet.getLong( "organisationunitid" ) );
     }
     
     @Override
